Breaking topics out by observation type and adding keys
Topics will now be created for the main observation types. If the topic does not exist when the subscriber starts up, it will be created automatically with # of partitions specified in kafka broker settings in server.properties

Will now receive the key in the unprocessed topics and pass that same key along to the transformed topic

Adding option to the read.me of --without-web so that multiple instances of cdip-routing can be started up without colliding on the web socket. Might be a way to configure this in the future but for now I dont think web as needed and this allows us to have multiple consumer instances